git commit -m"Implement feedback from LFA
- further place functions in helper file [Finishes #159423414]
