+++ server/routes/userRoutes.js
@@ -1,7 +1,7 @@
 import express from 'express';
-import userSignupController from '../controllers/userController/signup';
-import userLoginController from '../controllers/userController/login';
-import signupValidator from '../helpers/signupIndexValidator';
+import userSignupController from '../controllers/userController/Signup';
+import userLoginController from '../controllers/userController/Login';
+import { signupValidator } from '../helpers/helper';
 
 const router = express.Router();
